Add files via upload
Here is a Python project I made that I used as a fun game to test the user's knowledge of basic trivia questions. I wanted to learn Python programming and some of use its basic functionalities such as input/output, if/else, and boolean logic to create such a program.